Got uint block ids working (and other stuff)

+++ shaders/fragment.glsl
@@ -4,7 +4,7 @@ precision highp float;
 precision lowp sampler2DArray;
 
 in vec2 v_TexCoordBlock;
-in flat float blockID;
+in flat uint blockID;
 in float corner_lighting;
 
 uniform sampler2DArray u_Texture;
@@ -16,6 +16,6 @@ void main( ) {
     if ( texColor.a == 0.0 ) {
         discard;
     }
-    float corner_light = ( 3.0f - corner_lighting ) / 3.0;
+    float corner_light = ( 3.0 - corner_lighting ) / 3.0;
     color = texColor * vec4( corner_light, corner_light, corner_light, 1 );
 }
